Refactored the way how Views and templates are handled.
In the previous version of LJCore, the controller would determine what View class to fetch depending on the response format needed (JSON, HTML, XML etc.). This is not really intuitive, because it creates multiple View classes in subfolers for one Action.

Therefore, the controller has been refacterod in a way that always only 1 View class gets called, no matter what the response format is. That way, Action to View are one-to-one instead of one-to-many. If other response formats are needed, only the mustache templates sent back will be different.

I will make a more elaborate guide for using this model when I have time for it.